fix(web): 🐛 Complete visual story safety validation
